url/URL: Exchange fnv64 for fnv32 for gcc -m32 by default.

+++ url/URL.hpp
@@ -23,7 +23,7 @@ class URL
 {
     public:
         /* Dependent typedefs */
-        typedef uint64_t hash_t;
+        typedef uint32_t hash_t;
         typedef Iterator iterator;
         typedef TreeIterator tree_iterator;
